migrate to latest packaging
