--- src/main/kotlin/pw/forst/katlib/BooleanExtensions.kt
@@ -1,17 +1,0 @@
-package pw.forst.katlib
-
-/**
- * Executes [block] iff this (result of previous method) is true. Returns given Boolean.
- * */
-inline fun Boolean.whenTrue(block: () -> Unit): Boolean {
-    if (this) block()
-    return this
-}
-
-/**
- * Executes [block] iff this (result of previous method) is false. Returns given Boolean.
- * */
-inline fun Boolean.whenFalse(block: () -> Unit): Boolean {
-    if (!this) block()
-    return this
-}
