+++ src/test/resources/dev/forst/katlib/classlevel.properties
